Packages classes renamed to Plugins
