+++ tests/Support/LaravelPackages.php
@@ -2,12 +2,12 @@
 
 namespace Tests\Support;
 
-use Helldar\LaravelLangPublisher\Packages\Cashier;
-use Helldar\LaravelLangPublisher\Packages\Fortify;
-use Helldar\LaravelLangPublisher\Packages\Jetstream;
-use Helldar\LaravelLangPublisher\Packages\Nova;
-use Helldar\LaravelLangPublisher\Packages\SparkPaddle;
-use Helldar\LaravelLangPublisher\Packages\SparkStripe;
+use Helldar\LaravelLangPublisher\Plugins\Cashier;
+use Helldar\LaravelLangPublisher\Plugins\Fortify;
+use Helldar\LaravelLangPublisher\Plugins\Jetstream;
+use Helldar\LaravelLangPublisher\Plugins\Nova;
+use Helldar\LaravelLangPublisher\Plugins\SparkPaddle;
+use Helldar\LaravelLangPublisher\Plugins\SparkStripe;
 use Tests\TestCase;
 
 final class LaravelPackages extends TestCase
